hammer_referrals/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 
flake8==7.0.0
idna==3.7
mccabe==0.7.0
phonenumbers==8.13.35
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.5.0
urllib3==2.2.1
